refactor(search): use trending contract data in search (#3652)
<!-- start pr-codex -->

## PR-Codex overview
This PR updates time range handling and contract fetching in the trending feature.

### Detailed summary
- Updated time range handling in `TimeRangeSwitcher`
- Moved `TimeRange` type to `lib/search`
- Refactored contract fetching logic to `fetchTopContracts`
- Added new types and functions related to trending contracts
- Updated imports and removed redundant code in `trending-table`

> The following files were skipped due to too many changes: `apps/dashboard/src/app/trending/components/trending-table.tsx`, `apps/dashboard/src/components/cmd-k-search/index.tsx`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
