+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/page.tsx
@@ -1,10 +1,8 @@
 import { ChevronRight, ShieldCheckIcon } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
-import {
-  type TimeRange,
-  TrendingContractSection,
-} from "../../../trending/components/trending-table";
+import type { TimeRange } from "../../../../lib/search";
+import { TrendingContractSection } from "../../../trending/components/trending-table";
 import { getChain } from "../utils";
 import { InfoCard } from "./components/server/info-card";
 import twPublisherImage from "./tw-publisher.png";
