--- apps/dashboard/src/app/trending/components/icons.tsx
@@ -1,164 +1,0 @@
-export function getMedalIcon(index: number) {
-  switch (index) {
-    case 0:
-      return first;
-    case 1:
-      return second;
-    case 2:
-      return third;
-    default:
-      return null;
-  }
-}
-
-const first = (
-  <svg
-    width="23"
-    height="30"
-    viewBox="0 0 24 33"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    role="presentation"
-  >
-    <path
-      d="M0.47699 3.01535V24.3137C0.47699 25.016 0.84553 25.6668 1.4478 26.0281L11.0036 31.7593C11.6416 32.142 12.4393 32.1387 13.0742 31.7508L22.6324 25.91C23.2267 25.5468 23.5891 24.9006 23.5891 24.2042V3.01535C23.5891 1.91134 22.6941 1.01636 21.5901 1.01636H2.47599C1.37197 1.01636 0.47699 1.91134 0.47699 3.01535Z"
-      fill="url(#paint0_linear_20_1098)"
-      stroke="black"
-      stroke-width="0.399799"
-    />
-    <path
-      opacity="0.7"
-      d="M3.01984 4.33134V23.1563C3.01984 23.4307 3.16053 23.6859 3.39253 23.8324L11.6115 29.0219C11.8743 29.1879 12.2095 29.1865 12.471 29.0183L20.6847 23.7343C20.9135 23.5871 21.0517 23.3338 21.0517 23.0618V4.33134C21.0517 3.88973 20.6937 3.53174 20.2521 3.53174H3.81943C3.37783 3.53174 3.01984 3.88973 3.01984 4.33134Z"
-      fill="url(#paint1_linear_20_1098)"
-    />
-    <path
-      d="M14.2625 20.6172H11.529V11.4058C10.9674 11.6679 9.91891 11.8926 9.30107 11.9488V9.94548C10.4806 9.72081 11.6601 9.06553 12.3528 8.27919H14.2625V20.6172Z"
-      fill="#FFFDF3"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear_20_1098"
-        x1="12.0331"
-        y1="1.01636"
-        x2="12.0331"
-        y2="51.2802"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#F7E59D" />
-        <stop offset="1" stop-color="#FDC114" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_20_1098"
-        x1="12.0358"
-        y1="-1.09116"
-        x2="12.0358"
-        y2="29.294"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#F1B21A" />
-        <stop offset="1" stop-color="#FE9A06" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-const second = (
-  <svg
-    width="23"
-    height="30"
-    viewBox="0 0 24 33"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    role="presentation"
-  >
-    <path
-      d="M0.678894 3.01535V24.3137C0.678894 25.016 1.04743 25.6668 1.64971 26.0281L11.2055 31.7593C11.8435 32.142 12.6412 32.1387 13.2761 31.7508L22.8344 25.91C23.4286 25.5468 23.791 24.9006 23.791 24.2042V3.01535C23.791 1.91134 22.896 1.01636 21.792 1.01636H2.67789C1.57388 1.01636 0.678894 1.91134 0.678894 3.01535Z"
-      fill="url(#paint0_linear_20_1099)"
-      stroke="black"
-      stroke-width="0.399799"
-    />
-    <path
-      d="M3.22174 4.33134V23.1563C3.22174 23.4307 3.36243 23.6859 3.59444 23.8324L11.8134 29.0219C12.0762 29.1879 12.4114 29.1865 12.6729 29.0183L20.8867 23.7343C21.1154 23.5871 21.2536 23.3338 21.2536 23.0618V4.33134C21.2536 3.88973 20.8956 3.53174 20.454 3.53174H4.02134C3.57973 3.53174 3.22174 3.88973 3.22174 4.33134Z"
-      fill="url(#paint1_linear_20_1099)"
-    />
-    <path
-      d="M16.6465 20.6173H7.94061V20.4488C7.94061 18.2957 8.89545 16.5546 11.5727 14.6075C13.3139 13.3343 13.8381 12.9037 13.8381 11.8178C13.8381 10.7507 13.3139 10.1515 12.2842 10.1515C11.0672 10.1515 10.6366 10.9004 10.4681 11.9676H7.88445C8.05295 9.85199 9.4384 8.07337 12.2842 8.07337C15.4483 8.07337 16.5341 9.96432 16.5341 11.7991C16.5341 13.5964 15.7478 14.4577 13.4824 16.0491C11.9659 17.1162 11.3481 17.8277 11.1421 18.3894H16.9835L16.6465 20.6173Z"
-      fill="#F2F2F2"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear_20_1099"
-        x1="12.235"
-        y1="1.01636"
-        x2="12.235"
-        y2="51.2802"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#D9D9D9" />
-        <stop offset="1" stop-color="#597074" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_20_1099"
-        x1="12.2377"
-        y1="-5.84884"
-        x2="12.2377"
-        y2="29.294"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#D9D9D9" />
-        <stop offset="1" stop-color="#546F75" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-const third = (
-  <svg
-    width="23"
-    height="30"
-    viewBox="0 0 25 33"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    role="presentation"
-  >
-    <path
-      d="M0.781616 3.01572V24.3141C0.781616 25.0164 1.15016 25.6672 1.75243 26.0284L11.3083 31.7597C11.9463 32.1423 12.744 32.139 13.3788 31.7511L22.9371 25.9103C23.5313 25.5472 23.8937 24.901 23.8937 24.2046V3.01572C23.8937 1.9117 22.9988 1.01672 21.8947 1.01672H2.78061C1.6766 1.01672 0.781616 1.9117 0.781616 3.01572Z"
-      fill="url(#paint0_linear_20_1100)"
-      stroke="black"
-      stroke-width="0.399799"
-    />
-    <path
-      opacity="0.7"
-      d="M3.32446 4.33158V23.1565C3.32446 23.4309 3.46515 23.6862 3.69716 23.8327L11.9161 29.0222C12.1789 29.1882 12.5142 29.1867 12.7756 29.0186L20.9894 23.7345C21.2181 23.5874 21.3564 23.334 21.3564 23.062V4.33158C21.3564 3.88997 20.9984 3.53198 20.5568 3.53198H4.12406C3.68246 3.53198 3.32446 3.88997 3.32446 4.33158Z"
-      fill="url(#paint1_linear_20_1100)"
-    />
-    <path
-      d="M10.553 11.4624H8.00677C8.17527 9.94584 9.24244 8.07362 12.3316 8.07362C15.1774 8.07362 16.4505 9.68373 16.4505 11.4249C16.4505 12.904 15.5893 13.6154 14.9153 13.9337V13.9899C15.7204 14.2707 16.9186 15.0383 16.9186 16.948C16.9186 18.7266 15.8514 20.8422 12.3129 20.8422C8.99905 20.8422 7.87571 18.8577 7.76338 17.2288H10.3658C10.5343 18.0713 11.1147 18.7828 12.3878 18.7828C13.6235 18.7828 14.2413 17.9964 14.2413 16.9105C14.2413 15.7685 13.6422 15.0758 12.2193 15.0758H11.0959V13.0163H12.0882C13.3052 13.0163 13.8107 12.6044 13.8107 11.5747C13.8107 10.7322 13.3052 10.1518 12.238 10.1518C10.9836 10.1518 10.6841 10.8632 10.553 11.4624Z"
-      fill="#FFF3E3"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear_20_1100"
-        x1="12.3377"
-        y1="1.01672"
-        x2="12.3377"
-        y2="51.2806"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#FFCD9A" />
-        <stop offset="1" stop-color="#DB914D" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_20_1100"
-        x1="12.3404"
-        y1="-1.09092"
-        x2="12.3404"
-        y2="29.2943"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#D19355" />
-        <stop offset="1" stop-color="#D17920" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
