+++ apps/dashboard/src/app/trending/components/time-range-switcher.client.tsx
@@ -4,7 +4,7 @@ import {
   ToggleGroup,
   ToggleGroupItem,
 } from "../../../@/components/ui/toggle-group";
-import type { TimeRange } from "./trending-table";
+import type { TimeRange } from "../../../lib/search";
 
 export function TimeRangeSwitcher() {
   const router = useRouter();
