+++ apps/dashboard/src/app/trending/components/trending-table.tsx
@@ -1,10 +1,5 @@
-import {
-  type ChainMetadata,
-  defineChain,
-  getChainMetadata,
-} from "thirdweb/chains";
+import type { ChainMetadata } from "thirdweb/chains";
 import { shortenAddress } from "thirdweb/utils";
-import invariant from "tiny-invariant";
 import { ChainIcon } from "../../(dashboard)/(chain)/components/server/chain-icon";
 import {
   Table,
@@ -15,177 +10,13 @@ import {
   TableHeader,
   TableRow,
 } from "../../../@/components/ui/table";
+import {
+  type TrendingContract,
+  type TrendingContractProps,
+  fetchTopContracts,
+} from "../../../lib/search";
 import { TablePagination } from "./pagination.client";
 import { TimeRangeSwitcher } from "./time-range-switcher.client";
-
-export type TimeRange = "month" | "week" | "day";
-
-type TrendingContractProps = {
-  chainId?: number;
-  perPage?: number;
-  page?: number;
-  timeRange?: TimeRange;
-};
-
-const typesenseApiKey =
-  process.env.NEXT_PUBLIC_TYPESENSE_CONTRACT_API_KEY || "";
-
-type TrendingContract = {
-  name: string;
-  chainMetadata: ChainMetadata;
-  contractAddress: string;
-  transactionCount: string;
-  transactionCountChange: number;
-  walletCount: string;
-  walletCountChange: number;
-  gasUsage: string;
-  gasUsageChange: number;
-  valueMoved: string;
-  valueMovedChange: number;
-  type?: string;
-};
-
-const DOMAIN = "https://search.thirdweb.com";
-const COLLECTION = "hot_contracts";
-const TYPESENSE_URL = `${DOMAIN}/collections/${COLLECTION}/documents/search`;
-
-export const getSearchQueryUrl = ({
-  page = 1,
-  perPage = 10,
-  query = "",
-  chainId,
-  timeRange = "day",
-}: {
-  page?: number;
-  perPage?: number;
-  query?: string;
-  chainId?: number;
-  timeRange?: TimeRange;
-}) => {
-  const baseUrl = new URL(TYPESENSE_URL);
-  baseUrl.searchParams.set("query_by", "name,chainId,contractAddress");
-  baseUrl.searchParams.set("q", query);
-  baseUrl.searchParams.set("page", page.toString());
-  baseUrl.searchParams.set("per_page", perPage.toString());
-  if (chainId) {
-    baseUrl.searchParams.set("filter_by", `chainId:=${chainId}`);
-  }
-  baseUrl.searchParams.set(
-    "sort_by",
-    timeRange === "month"
-      ? "last30Days.transactionCount:desc"
-      : timeRange === "week"
-        ? "last7Days.transactionCount:desc"
-        : "last24Hr.transactionCount:desc",
-  );
-  return baseUrl.toString();
-};
-
-export async function fetchTopContracts(args?: {
-  chainId?: number;
-  query?: string;
-  perPage?: number;
-  page?: number;
-  timeRange?: TimeRange;
-}) {
-  invariant(typesenseApiKey, "No typesense api key");
-  const res = await fetch(
-    getSearchQueryUrl({
-      perPage: args?.perPage || 10,
-      page: args?.page || 1,
-      chainId: args?.chainId,
-      query: args?.query,
-      timeRange: args?.timeRange,
-    }),
-    {
-      headers: {
-        "x-typesense-api-key": typesenseApiKey,
-      },
-      next: { revalidate: 120 },
-    },
-  );
-  const result = await res.json();
-  if (!result.hits) return [];
-  return (await Promise.all(
-    // biome-ignore lint/suspicious/noExplicitAny: FIXME
-    result.hits.map(async (hit: any) => {
-      const document = hit.document;
-      const chainMetadata = await getChainMetadata(
-        defineChain(Number(document.chainId)),
-      );
-      let name = document.name;
-      if (
-        !name ||
-        name === "Unimported Contract" ||
-        name === "UnknownContract"
-      ) {
-        name = await fetchContractName(
-          document.chainId,
-          document.contractAddress,
-        );
-      }
-
-      let timeData: {
-        walletCount: number;
-        walletCountChange: number;
-        transactionCount: number;
-        transactionCountChange: number;
-        gasUsage: number;
-        gasUsageChange: number;
-        totalValueMoved: number;
-        totalValueMovedChange: number;
-      };
-      switch (args?.timeRange) {
-        case "week":
-          timeData = document.last7Days;
-          break;
-        case "day":
-          timeData = document.last24Hr;
-          break;
-        case "month":
-          timeData = document.last30Days;
-          break;
-        default:
-          timeData = document.last24Hr;
-          break;
-      }
-
-      return {
-        name,
-        chainMetadata,
-        contractAddress: document.contractAddress,
-        transactionCount: formatNumber(timeData.transactionCount),
-        transactionCountChange: timeData.transactionCountChange,
-        walletCount: formatNumber(timeData.walletCount),
-        walletCountChange: timeData.walletCountChange,
-        gasUsage: `${formatNumber(timeData.gasUsage)} ${chainMetadata.nativeCurrency.symbol}`,
-        gasUsageChange: timeData.gasUsageChange,
-        valueMoved:
-          timeData.totalValueMoved !== 0
-            ? `${formatNumber(timeData.totalValueMoved)} ${chainMetadata.nativeCurrency.symbol}`
-            : "",
-        valueMovedChange: timeData.totalValueMovedChange,
-        type: (document.type as Array<string>)?.find((t) =>
-          t.startsWith("ERC"),
-        ),
-      } as TrendingContract;
-    }),
-  )) as TrendingContract[];
-}
-
-async function fetchContractName(chainId: number, contractAddress: string) {
-  const res = await fetch(
-    `https://contract.thirdweb.com/metadata/${chainId}/${contractAddress}`,
-  );
-  const json = await res.json();
-  if (json.error) {
-    return "";
-  }
-  const compilationTarget = json.settings.compilationTarget;
-  const targets = Object.keys(compilationTarget);
-  const name = compilationTarget[targets[0]];
-  return name;
-}
 
 export async function TrendingContractSection(props: TrendingContractProps) {
   const isLandingPage = props.chainId === undefined;
@@ -235,7 +66,15 @@ export async function TrendingContractSection(props: TrendingContractProps) {
           {topContracts.map((contract, index) => (
             <TableRow
               key={`${contract.chainMetadata.chainId}${contract.contractAddress}${index}`}
-              className={`${isLandingPage && index + firstIndex < 3 ? `bg-orange-${300 + index * 100}/20` : isLandingPage && "bg-black/50"}  ${isLandingPage && "border-white !border-x-[1px]"}`}
+              className={`${isLandingPage && "border-white !border-x-[1px]"}`}
+              style={{
+                backgroundColor:
+                  isLandingPage && index + firstIndex < 3
+                    ? `rgb(200, 72, 0, ${0.3 - index * 0.05})`
+                    : isLandingPage
+                      ? "rgba(0, 0, 0, 0.5)"
+                      : undefined,
+              }}
             >
               <TableCell className="text-left">
                 {index + 1 + firstIndex}
@@ -334,22 +173,3 @@ function getContractName(contract: TrendingContract) {
   }
   return shortenAddress(contract.contractAddress);
 }
-
-const formatNumber = (num: number) => {
-  if (num >= 1000000) {
-    return `${(num / 1000000).toLocaleString(undefined, {
-      minimumFractionDigits: 1,
-      maximumFractionDigits: 1,
-    })}M`;
-  }
-  if (num >= 1000) {
-    return `${(num / 1000).toLocaleString(undefined, {
-      minimumFractionDigits: 1,
-      maximumFractionDigits: 1,
-    })}k`;
-  }
-  return num.toLocaleString(undefined, {
-    minimumFractionDigits: 0,
-    maximumFractionDigits: num < 10 ? 2 : 1,
-  });
-};
