+++ apps/dashboard/src/lib/search.ts
@@ -1,52 +1,201 @@
-export type SearchMode = "all" | "mainnet" | "testnet";
+import {
+  type ChainMetadata,
+  defineChain,
+  getChainMetadata,
+} from "thirdweb/chains";
+import invariant from "tiny-invariant";
 
-const TYPESENSE_URL =
-  "https://search.thirdweb.com/collections/contracts/documents/search";
+export type TimeRange = "month" | "week" | "day";
 
-export const getSearchQuery = ({
-  query,
-  walletAddress = "",
-  searchMode = "all",
+export type TrendingContractProps = {
+  chainId?: number;
+  perPage?: number;
+  page?: number;
+  timeRange?: TimeRange;
+};
+
+const typesenseApiKey =
+  process.env.NEXT_PUBLIC_TYPESENSE_CONTRACT_API_KEY || "";
+
+export type TrendingContract = {
+  name: string;
+  chainMetadata: ChainMetadata;
+  contractAddress: string;
+  transactionCount: string;
+  transactionCountChange: number;
+  walletCount: string;
+  walletCountChange: number;
+  gasUsage: string;
+  gasUsageChange: number;
+  valueMoved: string;
+  valueMovedChange: number;
+  type?: string;
+};
+
+const DOMAIN = "https://search.thirdweb.com";
+const COLLECTION = "hot_contracts";
+const TYPESENSE_URL = `${DOMAIN}/collections/${COLLECTION}/documents/search`;
+
+export const getSearchQueryUrl = ({
   page = 1,
   perPage = 10,
-  getAllOwnedByWallet = false,
-  filterByChains = [],
+  query = "",
+  chainId,
+  timeRange = "day",
 }: {
-  query: string;
-  walletAddress?: string;
-  searchMode: SearchMode;
   page?: number;
   perPage?: number;
-  getAllOwnedByWallet?: boolean;
-  filterByChains?: number[];
+  query?: string;
+  chainId?: number;
+  timeRange?: TimeRange;
 }) => {
   const baseUrl = new URL(TYPESENSE_URL);
+  baseUrl.searchParams.set("query_by", "name,chainId,contractAddress");
   baseUrl.searchParams.set("q", query);
-  baseUrl.searchParams.set(
-    "query_by",
-    "name,symbol,contract_address,deployer_address",
-  );
-  baseUrl.searchParams.set("query_by_weights", "3,3,2,1");
   baseUrl.searchParams.set("page", page.toString());
   baseUrl.searchParams.set("per_page", perPage.toString());
-  baseUrl.searchParams.set("exhaustive_search", "true");
+  const dateFilter =
+    timeRange === "month"
+      ? `updatedAt:> ${new Date().getTime() - 30 * 24 * 60 * 60 * 1000}`
+      : timeRange === "week"
+        ? `updatedAt:> ${new Date().getTime() - 7 * 24 * 60 * 60 * 1000}`
+        : `updatedAt:> ${new Date().getTime() - 1 * 24 * 60 * 60 * 1000}`;
+  baseUrl.searchParams.set(
+    "filter_by",
+    `${dateFilter}${chainId ? ` && chainId:=${chainId}` : ""}`,
+  );
   baseUrl.searchParams.set(
     "sort_by",
-    `testnet:asc${
-      walletAddress ? `,_eval(deployer_address:${walletAddress}):desc` : ""
-    }`,
+    timeRange === "month"
+      ? "last30Days.transactionCount:desc"
+      : timeRange === "week"
+        ? "last7Days.transactionCount:desc"
+        : "last24Hr.transactionCount:desc",
   );
-
-  if (searchMode === "mainnet") {
-    baseUrl.searchParams.set("filter_by", "testnet:false");
-  } else if (searchMode === "testnet") {
-    baseUrl.searchParams.set("filter_by", "testnet:true");
-  }
-  if (getAllOwnedByWallet) {
-    baseUrl.searchParams.set(
-      "filter_by",
-      `deployer_address:${walletAddress} && chain_id:[${filterByChains}]`,
-    );
-  }
   return baseUrl.toString();
+};
+
+export async function fetchTopContracts(args?: {
+  chainId?: number;
+  query?: string;
+  perPage?: number;
+  page?: number;
+  timeRange?: TimeRange;
+}) {
+  invariant(typesenseApiKey, "No typesense api key");
+  const res = await fetch(
+    getSearchQueryUrl({
+      perPage: args?.perPage || 10,
+      page: args?.page || 1,
+      chainId: args?.chainId,
+      query: args?.query,
+      timeRange: args?.timeRange,
+    }),
+    {
+      headers: {
+        "x-typesense-api-key": typesenseApiKey,
+      },
+      next: { revalidate: 120 },
+    },
+  );
+  const result = await res.json();
+  if (!result.hits) return [];
+  return (await Promise.all(
+    // biome-ignore lint/suspicious/noExplicitAny: FIXME
+    result.hits.map(async (hit: any) => {
+      const document = hit.document;
+      const chainMetadata = await getChainMetadata(
+        defineChain(Number(document.chainId)),
+      );
+      let name = document.name;
+      if (
+        !name ||
+        name === "Unimported Contract" ||
+        name === "UnknownContract"
+      ) {
+        name = await fetchContractName(
+          document.chainId,
+          document.contractAddress,
+        );
+      }
+
+      let timeData: {
+        walletCount: number;
+        walletCountChange: number;
+        transactionCount: number;
+        transactionCountChange: number;
+        gasUsage: number;
+        gasUsageChange: number;
+        totalValueMoved: number;
+        totalValueMovedChange: number;
+      };
+      switch (args?.timeRange) {
+        case "week":
+          timeData = document.last7Days;
+          break;
+        case "day":
+          timeData = document.last24Hr;
+          break;
+        case "month":
+          timeData = document.last30Days;
+          break;
+        default:
+          timeData = document.last24Hr;
+          break;
+      }
+
+      return {
+        name,
+        chainMetadata,
+        contractAddress: document.contractAddress,
+        transactionCount: formatNumber(timeData.transactionCount),
+        transactionCountChange: timeData.transactionCountChange,
+        walletCount: formatNumber(timeData.walletCount),
+        walletCountChange: timeData.walletCountChange,
+        gasUsage: `${formatNumber(timeData.gasUsage)} ${chainMetadata.nativeCurrency.symbol}`,
+        gasUsageChange: timeData.gasUsageChange,
+        valueMoved:
+          timeData.totalValueMoved !== 0
+            ? `${formatNumber(timeData.totalValueMoved)} ${chainMetadata.nativeCurrency.symbol}`
+            : "",
+        valueMovedChange: timeData.totalValueMovedChange,
+        type: (document.type as Array<string>)?.find((t) =>
+          t.startsWith("ERC"),
+        ),
+      } as TrendingContract;
+    }),
+  )) as TrendingContract[];
+}
+
+async function fetchContractName(chainId: number, contractAddress: string) {
+  const res = await fetch(
+    `https://contract.thirdweb.com/metadata/${chainId}/${contractAddress}`,
+  );
+  const json = await res.json();
+  if (json.error) {
+    return "";
+  }
+  const compilationTarget = json.settings.compilationTarget;
+  const targets = Object.keys(compilationTarget);
+  const name = compilationTarget[targets[0]];
+  return name;
+}
+
+const formatNumber = (num: number) => {
+  if (num >= 1000000) {
+    return `${(num / 1000000).toLocaleString(undefined, {
+      minimumFractionDigits: 1,
+      maximumFractionDigits: 1,
+    })}M`;
+  }
+  if (num >= 1000) {
+    return `${(num / 1000).toLocaleString(undefined, {
+      minimumFractionDigits: 1,
+      maximumFractionDigits: 1,
+    })}k`;
+  }
+  return num.toLocaleString(undefined, {
+    minimumFractionDigits: 0,
+    maximumFractionDigits: num < 10 ? 2 : 1,
+  });
 };
